backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.2
langchain-core>=0.3.15

# HuggingFace
huggingface-hub>=0.25.0
sentence-transformers==3.0.1

# Supabase
supabase==2.7.4
vecs==0.4.4

# Scraping
trafilatura==1.9.0
playwright==1.47.0

# Utilities
tiktoken==0.7.0
httpx==0.27.2
pydantic==2.8.2
